Here are four character editor.
The oldest one (I got from other website.) can't work with the newest one. So if you want that one, download it from the drive, open it in another unity project. Then, use spritesheetmaker.scene in the project to generate a sprite sheet to use. Well, it's not going to as good as these one, but it can work.